Move core tests
