+++ modules/core/arrow-core/build.gradle
@@ -7,6 +7,7 @@ dependencies {
     testCompileOnly project(':arrow-annotations-processor')
     testCompile "io.kotlintest:kotlintest:$kotlinTestVersion"
     testCompile project(':arrow-test')
+    testCompile project(':arrow-syntax')
 }
 
 apply from: rootProject.file('gradle/gradle-mvn-push.gradle')
