+++ modules/core/arrow-core/src/test/kotlin/arrow/core/EitherTest.kt
@@ -1,4 +1,4 @@
-package arrow.data
+package arrow.core
 
 import arrow.Kind
 import arrow.Kind2
